Move ChatbotDefinition to common models
