+++ src/rpft/parsers/creation/contentindexparser.py
@@ -11,6 +11,7 @@ from rpft.parsers.creation.campaignparser import CampaignParser
 from rpft.parsers.creation.contentindexrowmodel import ContentIndexRowModel
 from rpft.parsers.creation.flowparser import FlowParser
 from rpft.parsers.creation.global_templates import get_survey_reader
+from rpft.parsers.creation.models import ChatbotDefinition
 from rpft.parsers.creation.tagmatcher import TagMatcher
 from rpft.parsers.creation.surveyparser import Survey, SurveyParser
 from rpft.parsers.creation.triggerparser import TriggerParser
@@ -41,23 +42,6 @@ class DataSheet:
             "model": self.row_model.__name__,
             "rows": [content.dict() for content in self.rows.values()],
         }
-
-
-class ChatbotDefinition:
-    def __init__(self, flow_definitions, data_sheets, templates, surveys):
-        self.flow_definitions = flow_definitions
-        self.data_sheets = data_sheets
-        self.templates = templates
-        self.surveys = surveys
-
-    def get_data_sheet_rows(self, sheet_name):
-        return self.data_sheets[sheet_name].rows
-
-    def get_data_sheet_row(self, sheet_name, row_id):
-        return self.data_sheets[sheet_name].rows[row_id]
-
-    def get_template(self, name):
-        return self.templates[name]
 
 
 class ParserError(Exception):
